ipmi_server/requirements.txt
grpcio==1.10.0
protobuf==3.5.1
six==1.11.0
